requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
fastapi==0.108.0
h11==0.14.0
httptools==0.6.1
idna==3.6
pydantic==2.5.3
python-dotenv==1.0.0
python-decouple==3.8
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.32.0.post1
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
httpx~=0.26.0
motor~=3.3.2